don't trigger full reload when other cells execute
When a module is updated but not accepted by its importers, Vite triggers
a full page reload. This is wrong when the module is imported (so it's part of the module graph) but not actually used by the currently-viewed page; in that case the page is reloaded unnecessarily.

This is a corner case for ordinary Vite usage, where you're usually looking at just
one page and only importing its dependencies; but for Vitale each cell is an entry point,
so modules outside the dependencies of each entry point are loaded, and when they
change a full reload may be triggered.

We fix it by patching the HMR handler in the client, so it doesn't full-reload if the triggering file doesn't match the loaded URI.