don't execute paused cells when initiated from client
fixes #48. also clarified usages of `force`.

this is a little bit hairy because execution can be initiated from either client
(user runs a cell, switches cell focus, or clicks Run Dirty) or server
(external change triggers execution of dirty dependencies), and the server
doesn't know if a cell is paused; so we need to check in a few places.